Отрефакторит модуль с утилитами
